backend/requirements.txt
absl-py==2.2.2
accelerate==0.34.2
adagio==0.2.6
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-cors==0.8.1
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
autogluon==1.3.0
autogluon.common==1.3.0
autogluon.core==1.3.0
autogluon.features==1.3.0
autogluon.multimodal==1.3.0
autogluon.tabular==1.3.0
autogluon.timeseries==1.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
blis==0.7.11
boto3==1.38.10
botocore==1.38.10
cachetools==5.5.2
catalogue==2.0.10
catboost==1.2.8
category_encoders==2.7.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.6
colorlog==6.9.0
colour==0.1.5
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
coreforecast==0.0.12
cryptography==45.0.2
cycler==0.12.1
cymem==2.0.11
Cython==3.1.1
dash==3.0.4
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash-testing-stub==0.0.2
dataframe_api_compat==0.2.7
datasets==2.14.4
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.7
distlib==0.3.9
distributed==2025.5.0
ecdsa==0.19.1
einops==0.8.1
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.3.0
executing==2.2.0
fastai==2.7.19
fastapi==0.115.12
fastcore==1.7.29
fastdownload==0.0.7
fastjsonschema==2.21.1
fastprogress==1.0.3
filelock==3.18.0
Flask==3.0.3
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.6.0
fs==2.4.16
fsspec==2025.3.2
fugue==0.9.1
future==1.0.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
gluonts==0.16.1
google-api-core==2.24.2
google-auth==2.40.1
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.1
grpcio==1.71.0
h11==0.16.0
holidays==0.72
htmlmin==0.1.12
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
hyperopt==0.2.7
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.36.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.3.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaleido==0.2.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
lightgbm==4.5.0
lightning==2.5.1.post0
lightning-utilities==0.14.3
llvmlite==0.44.0
locket==1.0.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mlforecast==0.13.4
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.15
murmurhash==1.0.12
narwhals==1.38.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nlpaug==1.1.11
nltk==3.8.1
notebook==7.4.2
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvidia-ml-py3==7.352.0
omegaconf==2.2.3
opencensus==0.11.4
opencensus-context==0.1.3
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.5
openxlab==0.0.11
optuna==4.3.0
ordered-set==4.1.0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
patsy==1.0.1
pdf2image==1.17.0
pillow==11.2.1
platformdirs==4.3.7
plotly==5.20.0
plotly-resampler==0.10.0
pluggy==1.6.0
pmdarima==2.0.4
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.1.1
pure_eval==0.2.3
puremagic==1.29
py-spy==0.4.0
py4j==0.10.9.9
pyarrow==19.0.1
pyasn1==0.4.1
pyasn1-modules==0.2.5
pycaret==3.3.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyod==2.0.5
pyparsing==3.2.3
PySocks==1.7.1
pytesseract==0.3.10
pytest==8.4.0
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-lightning==2.5.1
pytorch-metric-learning==2.3.0
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.4.0
ray==2.39.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
ruff==0.11.10
s3transfer==0.12.0
safetensors==0.5.3
schemdraw==0.15
scikit-base==0.7.8
scikit-image==0.24.0
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
seqeval==1.2.2
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
sktime==0.26.0
slicer==0.0.7
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
sqlglot==26.17.1
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
statsforecast==1.7.8
statsmodels==0.14.4
streamlit==1.43.2
sympy==1.13.1
tabulate==0.9.0
tbats==1.1.3
tbb==2022.1.0
tblib==3.1.0
tcmlib==1.3.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.18.1
text-unidecode==1.3
thinc==8.2.5
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.3
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
toolz==0.12.1
torch==2.5.1
torchmetrics==1.2.1
torchvision==0.20.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triad==0.9.8
tsdownsample==0.1.4.1
typeguard==4.4.2
typer==0.15.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
uri-template==1.3.0
urllib3==1.26.20
urllib3-secure-extra==0.1.0
utilsforecast==0.2.4
uuid6==2024.7.10
uvicorn==0.34.2
virtualenv==20.31.1
waitress==3.0.2
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
widgetsnbextension==4.0.14
window_ops==0.0.15
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
xgboost==2.1.4
XlsxWriter==3.2.3
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.20.0
yellowbrick==1.5
zict==3.0.0
zipp==3.22.0
zope.event==5.0
zope.interface==7.2
